# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Area Backend API
  version: 1.0.0
  description: |-
    HTTP API exposed by the Area backend. All routes are mounted under `CONFIG.BASE_PATH`
    (default `/api`). Authentication uses bearer JWTs issued by `/auth/login`. Runner
    endpoints use a shared nonce/token pair carried in custom headers.
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by `/auth/login`
    RunnerJobId:
      type: apiKey
      in: header
      name: X-Runner-Job
      description: Runner job identifier returned when the workflow execution was created
    RunnerToken:
      type: apiKey
      in: header
      name: X-Runner-Token
      description: HMAC token used to authenticate runner -> backend callbacks
    PasswordResetToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token sent with password reset emails ("Bearer <token>")
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human readable error message
    MessageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
...
    UserCredential:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
    CredentialRecord:
      type: object
      required:
        - id
        - type
        - data
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        version:
          type: string
        description:
          type: string
          nullable: true
        data:
          type: object
          additionalProperties: true
          description: Field names only; values are redacted placeholders ("***")
    CredentialSaveRequest:
      type: object
      required:
        - id
        - type
        - data
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        version:
          type: string
        description:
          type: string
          nullable: true
        data:
          type: object
          additionalProperties: true
          description: Full credential payload to store (actual secret values)
    UserWorkflowSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
...
    UserLinkRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
          description: User identifier
    TeamLinkRequest:
      type: object
      required:
        - teamId
      properties:
        teamId:
          type: string
          description: Team identifier
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: integer
        enabled:
          type: boolean
        definition:
          type: object
          description: Workflow definition payload
        data:
          type: object
          additionalProperties: true
          description: Arbitrary workflow metadata payload
      additionalProperties: true
...
paths:
...
  /credentials/public:
    get:
      tags:
        - Credentials
      summary: List credentials with no ownership restrictions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Public credential list
          content:
            application/json:
              schema:
                type: object
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: '#/components/schemas/CredentialRecord'
        '401':
          description: Missing or invalid token
        '404':
          description: No credentials available
        '500':
          description: Internal error
  /credentials:
    get:
      tags:
        - Credentials
      summary: List credentials owned by the caller
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Credential list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserCredential'
        '404':
          description: Not found
        '500':
          description: Internal error
    post:
      tags:
        - Credentials
      summary: Create or update a credential definition
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialSaveRequest'
      responses:
        '201':
          description: Credential persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Missing fields or invalid payload
        '401':
          description: Missing or invalid token
        '500':
          description: Internal error
  /credentials/all:
    get:
      tags:
        - Credentials
      summary: List all credentials (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Credential list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserCredential'
        '404':
          description: Not found
        '500':
          description: Internal error
  /credentials/{id}:
    get:
      tags:
        - Credentials
      summary: Retrieve a credential by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credential payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not allowed to use the credential
        '404':
          description: Credential not found
        '500':
          description: Internal error
    put:
      tags:
        - Credentials
      summary: Update an existing credential definition
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialSaveRequest'
      responses:
        '200':
          description: Credential updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request or payload
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not credential owner or admin
        '404':
          description: Credential not found
        '500':
          description: Internal error
    delete:
      tags:
        - Credentials
      summary: Delete a credential definition
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credential removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not credential owner or admin
        '404':
          description: Credential not found
        '500':
          description: Internal error
  /credentials/{id}/users:
    post:
      tags:
        - Credentials
      summary: Grant a user access to a credential
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLinkRequest'
      responses:
        '200':
          description: Updated credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Credential or user not found
  /credentials/{id}/users/{userId}:
    delete:
      tags:
        - Credentials
      summary: Revoke a user's access to a credential
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Credential not found
  /credentials/{id}/owners:
    post:
      tags:
        - Credentials
      summary: Promote or add a credential owner
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLinkRequest'
      responses:
        '200':
          description: Updated credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Credential or user not found
  /credentials/{id}/owners/{userId}:
    delete:
      tags:
        - Credentials
      summary: Remove an owner from a credential
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Credential not found
  /credentials/{id}/user-teams:
    post:
      tags:
        - Credentials
      summary: Grant a team user-level access to a credential
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamLinkRequest'
      responses:
        '200':
          description: Updated credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Credential or team not found
  /credentials/{id}/user-teams/{teamId}:
    delete:
      tags:
        - Credentials
      summary: Revoke a team's user-level access to a credential
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Credential not found
  /credentials/{id}/owner-teams:
    post:
      tags:
        - Credentials
      summary: Grant a team owner-level access to a credential (promotes if necessary)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamLinkRequest'
      responses:
        '200':
          description: Updated credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Credential or team not found
  /credentials/{id}/owner-teams/{teamId}:
    delete:
      tags:
        - Credentials
      summary: Revoke a team's owner-level access to a credential
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: teamId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialRecord'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not an owner or admin
        '404':
          description: Credential not found
...
  /workflows:
    get:
      tags:
        - Workflows
      summary: List workflows the caller belongs to
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Workflow summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserWorkflowSummary'
        '404':
          description: Not found
        '500':
          description: Internal error
    post:
      tags:
        - Workflows
      security:
        - bearerAuth: []
      summary: Create or update a workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workflow'
      responses:
        '201':
          description: Workflow stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          description: Invalid workflow payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
...
  /workflows/{id}:
    get:
      tags:
        - Workflows
      security:
        - bearerAuth: []
      summary: Load a workflow
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Workflows
      security:
        - bearerAuth: []
      summary: Update a workflow
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workflow'
      responses:
        '200':
          description: Workflow updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          description: Invalid workflow payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not workflow owner or admin
        '404':
          description: Workflow not found
    delete:
      tags:
        - Workflows
      security:
        - bearerAuth: []
      summary: Delete a workflow
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request
        '401':
          description: Missing or invalid token
        '403':
          description: Caller is not workflow owner or admin
        '404':
          description: Workflow not found
...